+++ src/modules/__init__.py
@@ -1,3 +1,2 @@
 from .input import Key
 from .config import Config
-from .messagebox import showerror
