+++ src/utils/__init__.py
@@ -1,0 +1,2 @@
+from .logger import log
+from .messagebox import showerror
